+++ apps/docs/src/content/03-components/content/color/examples/default.tsx
@@ -1,0 +1,11 @@
+import {
+  Color,
+  Text,
+} from "@mittwald/flow-react-components";
+
+<Text>
+  <b>
+    Veröffentliche{" "}
+    <Color color="violet">eigene Features</Color> im mStudio
+  </b>
+</Text>;