+++ packages/components/src/integrations/react-hook-form/components/ActionGroupWrapper/SubmitButtonStateProvider.tsx
@@ -3,8 +3,6 @@ import React from "react";
 import { ActionStateContext } from "@/components/Action/ActionStateContext";
 import { useFormContext } from "@/integrations/react-hook-form/components/context/formContext";
 import { useFormState } from "react-hook-form";
-import type { PropsContext } from "@/lib/propsContext";
-import { dynamic, PropsContextProvider } from "@/lib/propsContext";
 
 interface Props extends PropsWithChildren {
   isAsyncSubmit: MutableRefObject<boolean>;
@@ -39,19 +37,9 @@ const ActionStateContextWrapper: FC<Props> = (props) => {
 export const SubmitButtonStateProvider: FC<Props> = (props) => {
   const { children, isAsyncSubmit } = props;
 
-  const propsContext: PropsContext = {
-    ActionGroup: {
-      children: dynamic((p) => (
-        <ActionStateContextWrapper isAsyncSubmit={isAsyncSubmit}>
-          {p.children}
-        </ActionStateContextWrapper>
-      )),
-    },
-  };
-
   return (
-    <PropsContextProvider props={propsContext} mergeInParentContext>
+    <ActionStateContextWrapper isAsyncSubmit={isAsyncSubmit}>
       {children}
-    </PropsContextProvider>
+    </ActionStateContextWrapper>
   );
 };