+++ packages/ext-bridge/src/react/hooks/useExtBridge.ts
@@ -1,0 +1,7 @@
+import { usePromise } from "@mittwald/react-use-promise";
+
+export const useExtBridge = () =>
+  usePromise(async () => {
+    await globalThis.mwExtBridge.ready;
+    return globalThis.mwExtBridge;
+  }, []);