+++ CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.105](https://github.com/mittwald/flow/compare/0.2.0-alpha.104...0.2.0-alpha.105) (2025-03-21)
+
+### Features
+
+* **Button:** support async submit without action ([#1303](https://github.com/mittwald/flow/issues/1303)) ([38d9e9f](https://github.com/mittwald/flow/commit/38d9e9feb99b0785a699aa916971aaa2bbaff8ca))
+
+# [0.2.0-alpha.104](https://github.com/mittwald/flow/compare/0.2.0-alpha.103...0.2.0-alpha.104) (2025-03-21)
+
+### Features
+
+* **AccentBox:** support green color ([#1289](https://github.com/mittwald/flow/issues/1289)) ([cae2804](https://github.com/mittwald/flow/commit/cae280437a46aea0aa0fe5047b636fb88c808601))
+
+# [0.2.0-alpha.103](https://github.com/mittwald/flow/compare/0.2.0-alpha.102...0.2.0-alpha.103) (2025-03-20)
+
+### Features
+
+* **ExtBridge:** add Extension Bridge ([69bbcd5](https://github.com/mittwald/flow/commit/69bbcd50d5df33a3d54bfa1534fc2a84cdcf0a9f))
+
+# [0.2.0-alpha.102](https://github.com/mittwald/flow/compare/0.2.0-alpha.101...0.2.0-alpha.102) (2025-03-20)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.101](https://github.com/mittwald/flow/compare/0.2.0-alpha.100...0.2.0-alpha.101) (2025-03-20)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.100](https://github.com/mittwald/flow/compare/0.2.0-alpha.99...0.2.0-alpha.100) (2025-03-20)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
 # [0.2.0-alpha.99](https://github.com/mittwald/flow/compare/0.2.0-alpha.98...0.2.0-alpha.99) (2025-03-19)
 
 ### Features